protect default rank from being deleted in command

+++ common/src/main/java/com/guflan/kingdomcraft/common/commands/management/ranks/RanksDeleteCommand.java
@@ -58,6 +58,11 @@ public class RanksDeleteCommand extends CommandBase {
             return;
         }
 
+        if ( kingdom.getDefaultRank().equals(rank) ) {
+            kdc.getMessageManager().send(sender, "cmdRanksDeleteDefault");
+            return;
+        }
+
         kdc.deleteAsync(rank);
         kdc.getMessageManager().send(sender, "cmdRanksDelete", rank.getName());
     }
